Add files via upload
In this version, a .txt file containing a list of number. The first two row of this list are referring two the images found in the file stim, the last and third row is going to be used for showing those pairs at three different lags (1, 2 and 7). This list is uploaded into an numpy array and then shuffled and split into 3 (or the number of blocks). This assure that each pair is shown twich, but also that each pairs is gonna be represented into each lags of interest. 

An if condition as also been added to make sure that no stimulus (T1 or T2) is shown twice in a row.

Not a definitive version.