Add files via upload
Completed log information. The position of T1 is now randomized, in regard to the lag of T2. No longer require a .txt to load a block into it, now uses a function that return a matrices where T1, T2 and lags are found for each pairs at each row. An option can be uncommented if one would like to check if the content of their log match the matrice created.